Bulk export map info bug fixed Mystery water bug fixed Can now use scrolls and potions when mounted Renamed global (general) buff to "Mirian Buff" BoneHammer and SewingGoods functionality added.
BoneHammer needs to be itemtype "nothing", and shape 1
SewingGoods needs to be itemtype "nothing", and shape 2

To use, place item you want to repair in your bag. Click the hammer/sewing item, then hold CTRL and click the item you want to repair.

It will repair the item with 100% success, and increase its dura that of what the repairing item had.